+++ tests/library_test.py
@@ -78,6 +78,25 @@ def sample_library():
     return lib
 
 
+def test_gds_info(tmpdir, sample_library):
+    fname = str(tmpdir.join("test.gds"))
+    sample_library.write_gds(fname, max_points=20)
+    info = gdstk.gds_info(fname)
+    assert set(info["cell_names"]) == {
+        "gl_rw_gds_1",
+        "gl_rw_gds_2",
+        "gl_rw_gds_3",
+        "gl_rw_gds_4",
+    }
+    assert info["layers_and_datatypes"] == {(0, 0), (2, 4)}
+    assert info["layers_and_texttypes"] == {(5, 6)}
+    assert info["num_polygons"] == 3
+    assert info["num_paths"] == 0
+    assert info["num_references"] == 2
+    assert info["num_labels"] == 1
+    assert info["unit"] == 2e-3
+    assert info["precision"] == 1e-5
+
 def test_rw_gds(tmpdir, sample_library):
     fname = str(tmpdir.join("test.gds"))
     sample_library.write_gds(fname, max_points=20)
@@ -188,6 +207,59 @@ def test_rw_gds_filter(tmpdir, sample_library):
     assert c.references[0].repetition.rows == 3
     assert c.references[0].repetition.v1 == (-2.0, 0.0)
     assert c.references[0].repetition.v2 == (0.0, 8.0)
+
+
+# def test_rw_oas_filter(tmpdir, sample_library):
+#     fname = str(tmpdir.join("test.oas"))
+#     sample_library.write_oas(fname)
+#     library = gdstk.read_oas(fname, unit=1e-3, filter={(0, 0)})
+
+#     assert library.name == "LIB"
+#     assert len(library.cells) == 4
+#     cells = {c.name: c for c in library.cells}
+#     assert set(cells.keys()) == {
+#         "gl_rw_gds_1",
+#         "gl_rw_gds_2",
+#         "gl_rw_gds_3",
+#         "gl_rw_gds_4",
+#     }
+#     c = cells["gl_rw_gds_1"]
+#     assert len(c.polygons) == 0
+#     assert len(c.labels) == 1
+#     assert c.labels[0].text == "label"
+#     assert c.labels[0].origin[0] == 2 and c.labels[0].origin[1] == -2
+#     assert c.labels[0].anchor == "sw"
+#     assert c.labels[0].rotation == 0
+#     assert c.labels[0].magnification == 1
+#     assert c.labels[0].x_reflection == False
+#     assert c.labels[0].layer == 5
+#     assert c.labels[0].texttype == 6
+
+#     c = cells["gl_rw_gds_2"]
+#     assert len(c.polygons) == 1
+#     assert isinstance(c.polygons[0], gdstk.Polygon)
+
+#     c = cells["gl_rw_gds_3"]
+#     assert len(c.references) == 1
+#     assert isinstance(c.references[0], gdstk.Reference)
+#     assert c.references[0].cell == cells["gl_rw_gds_1"]
+#     assert c.references[0].origin[0] == 0 and c.references[0].origin[1] == 2
+#     assert c.references[0].rotation == -90
+#     assert c.references[0].magnification == 2
+#     assert c.references[0].x_reflection == True
+
+#     c = cells["gl_rw_gds_4"]
+#     assert len(c.references) == 1
+#     assert isinstance(c.references[0], gdstk.Reference)
+#     assert c.references[0].cell == cells["gl_rw_gds_2"]
+#     assert c.references[0].origin[0] == -2 and c.references[0].origin[1] == -4
+#     assert c.references[0].rotation == numpy.pi
+#     assert c.references[0].magnification == 0.5
+#     assert c.references[0].x_reflection == True
+#     assert c.references[0].repetition.columns == 2
+#     assert c.references[0].repetition.rows == 3
+#     assert c.references[0].repetition.v1 == (-2.0, 0.0)
+#     assert c.references[0].repetition.v2 == (0.0, 8.0)
 
 
 def test_rw_oas(tmpdir, sample_library):